refactor: Mejora la estructura de carpetas del proyecto
