+++ components/ui/dropdown-menu.tsx
@@ -5,7 +5,7 @@ import { Check, ChevronRight, Circle } from 'lucide-react';
 
 import * as React from 'react';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib/utils/cn';
 
 function DropdownMenu({
   ...props
