+++ components/ui/label.tsx
@@ -4,7 +4,7 @@ import * as LabelPrimitive from "@radix-ui/react-label"
 
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/utils/cn"
 
 function Label({
   className,
