+++ components/ui/tabs.tsx
@@ -4,7 +4,7 @@ import * as TabsPrimitive from "@radix-ui/react-tabs"
 
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/utils/cn"
 
 function Tabs({
   className,
