+++ domain/services/getStargazers.ts
@@ -1,4 +1,4 @@
-import { githubClient } from '@/lib/githubClient';
+import { githubClient } from '@/lib/github/githubClient';
 
 export async function getStargazers() {
   const { data } = await githubClient.request('GET /repos/{owner}/{repo}', {
