+++ lib/auth/auth.ts
