+++ lib/utils/numberFormat.ts
